+++ apps/GatewayApi/Controllers/TransactionController.cs
@@ -87,6 +87,12 @@ public sealed class TransactionController : ControllerBase
         return await _transactionHandler.Construction(token);
     }
 
+    [HttpPost("subintent-status")]
+    public async Task<GatewayModel.TransactionSubintentStatusResponse> Status(GatewayModel.TransactionSubintentStatusRequest request, CancellationToken token)
+    {
+        return await _transactionHandler.SubintentStatus(request, token);
+    }
+
     [HttpPost("status")]
     public async Task<GatewayModel.TransactionStatusResponse> Status(GatewayModel.TransactionStatusRequest request, CancellationToken token)
     {
@@ -103,6 +109,12 @@ public sealed class TransactionController : ControllerBase
     public async Task<GatewayModel.TransactionPreviewResponse> Preview(GatewayModel.TransactionPreviewRequest request, CancellationToken token)
     {
         return await _transactionHandler.Preview(request, token);
+    }
+
+    [HttpPost("preview-v2")]
+    public async Task<GatewayModel.TransactionPreviewV2Response> PreviewV2(GatewayModel.TransactionPreviewV2Request request, CancellationToken token)
+    {
+        return await _transactionHandler.PreviewV2(request, token);
     }
 
     [HttpPost("submit")]