+++ src/RadixDlt.NetworkGateway.GatewayApi/Handlers/ITransactionHandler.cs
@@ -74,9 +74,13 @@ public interface ITransactionHandler
 
     Task<GatewayModel.TransactionStatusResponse> Status(GatewayModel.TransactionStatusRequest request, CancellationToken token = default);
 
+    Task<GatewayModel.TransactionSubintentStatusResponse> SubintentStatus(GatewayModel.TransactionSubintentStatusRequest request, CancellationToken token = default);
+
     Task<GatewayModel.TransactionCommittedDetailsResponse> CommittedDetails(GatewayModel.TransactionCommittedDetailsRequest request, CancellationToken token = default);
 
     Task<GatewayModel.TransactionPreviewResponse> Preview(GatewayModel.TransactionPreviewRequest request, CancellationToken token = default);
+
+    Task<GatewayModel.TransactionPreviewV2Response> PreviewV2(GatewayModel.TransactionPreviewV2Request request, CancellationToken token = default);
 
     Task<GatewayModel.TransactionSubmitResponse> Submit(GatewayModel.TransactionSubmitRequest request, CancellationToken token = default);