+++ src/RadixDlt.NetworkGateway.GatewayApi/Services/ISubmissionTrackingService.cs
@@ -68,7 +68,6 @@ using RadixDlt.NetworkGateway.Abstractions.CoreCommunications;
 using System;
 using System.Threading;
 using System.Threading.Tasks;
-using ToolkitModel = RadixEngineToolkit;
 
 namespace RadixDlt.NetworkGateway.GatewayApi.Services;
 
@@ -79,7 +78,7 @@ public interface ISubmissionTrackingService
         string networkName,
         string nodeName,
         PendingTransactionHandlingConfig handlingConfig,
-        ToolkitModel.NotarizedTransaction notarizedTransaction,
+        ParsedTransactionData parsedTransactionData,
         byte[] notarizedTransactionBytes,
         TimeSpan submissionTimeout,
         long currentEpoch,