+++ src/RadixDlt.NetworkGateway.PostgresIntegration/CommonDbContext.cs
@@ -87,6 +87,10 @@ internal abstract class CommonDbContext : DbContext
 
     public DbSet<LedgerTransactionMarker> LedgerTransactionMarkers => Set<LedgerTransactionMarker>();
 
+    public DbSet<LedgerTransactionSubintentData> LedgerTransactionSubintentData => Set<LedgerTransactionSubintentData>();
+
+    public DbSet<LedgerFinalizedSubintent> LedgerFinalizedSubintents => Set<LedgerFinalizedSubintent>();
+
     public DbSet<PendingTransaction> PendingTransactions => Set<PendingTransaction>();
 
     public DbSet<Entity> Entities => Set<Entity>();
@@ -243,13 +247,14 @@ internal abstract class CommonDbContext : DbContext
         modelBuilder
             .Entity<LedgerTransaction>()
             .HasDiscriminator<LedgerTransactionType>(DiscriminatorColumnName)
-            .HasValue<UserLedgerTransaction>(LedgerTransactionType.User)
+            .HasValue<UserLedgerTransactionV1>(LedgerTransactionType.User)
+            .HasValue<UserLedgerTransactionV2>(LedgerTransactionType.UserV2)
             .HasValue<RoundUpdateLedgerTransaction>(LedgerTransactionType.RoundUpdate)
             .HasValue<GenesisLedgerTransaction>(LedgerTransactionType.Genesis)
             .HasValue<FlashLedgerTransaction>(LedgerTransactionType.Flash);
 
         modelBuilder
-            .Entity<UserLedgerTransaction>()
+            .Entity<BaseUserLedgerTransaction>()
             .HasIndex(lt => lt.IntentHash)
             .HasMethod("hash")
             .HasFilter("intent_hash IS NOT NULL");