+++ src/RadixDlt.NetworkGateway.PostgresIntegration/LedgerExtension/Processors/LedgerTransactionMarkers/ManifestProcessor.cs
@@ -98,10 +98,34 @@ internal class ManifestProcessor : ITransactionMarkerProcessor, ITransactionScan
         {
             var coreInstructions = userLedgerTransaction.NotarizedTransaction.SignedIntent.Intent.Instructions;
             var coreBlobs = userLedgerTransaction.NotarizedTransaction.SignedIntent.Intent.BlobsHex;
-            using var manifestInstructions = ToolkitModel.Instructions.FromString(coreInstructions, _networkConfiguration.Id);
-            using var toolkitManifest = new ToolkitModel.TransactionManifest(manifestInstructions, coreBlobs.Values.Select(x => x.ConvertFromHex()).ToArray());
+            using var manifestInstructions = ToolkitModel.InstructionsV1.FromString(coreInstructions, _networkConfiguration.Id);
+            using var toolkitManifest = new ToolkitModel.TransactionManifestV1(manifestInstructions, coreBlobs.Values.Select(x => x.ConvertFromHex()).ToArray());
 
-            AnalyzeManifestClasses(toolkitManifest, stateVersion);
+            AnalyzeManifestV1Classes(toolkitManifest, stateVersion);
+
+            if (transaction.Receipt.Status == CoreModel.TransactionStatus.Succeeded)
+            {
+                var extractedAddresses = ManifestAddressesExtractor.ExtractAddresses(toolkitManifest, _networkConfiguration.Id);
+
+                foreach (var address in extractedAddresses.All())
+                {
+                    _referencedEntities.MarkSeenAddress(address);
+                }
+
+                _manifestExtractedAddresses.Add(stateVersion, extractedAddresses);
+            }
+        }
+        else if (transaction.LedgerTransaction is CoreModel.UserLedgerTransactionV2 userLedgerTransactionV2)
+        {
+            var coreInstructions = userLedgerTransactionV2.NotarizedTransaction.SignedTransactionIntent.TransactionIntent.RootIntentCore.Instructions;
+            var coreBlobs = userLedgerTransactionV2.NotarizedTransaction.SignedTransactionIntent.TransactionIntent.RootIntentCore.BlobsHex;
+
+            var children = userLedgerTransactionV2.NotarizedTransaction.SignedTransactionIntent.TransactionIntent.NonRootSubintents
+                .Select(x => new ToolkitModel.Hash(Convert.FromHexString(x.Hash))).ToArray();
+            using var manifestInstructions = ToolkitModel.InstructionsV2.FromString(coreInstructions, _networkConfiguration.Id);
+            using var toolkitManifest = new ToolkitModel.TransactionManifestV2(manifestInstructions, coreBlobs.Values.Select(x => x.ConvertFromHex()).ToArray(), children);
+
+            AnalyzeManifestV2Classes(toolkitManifest, stateVersion);
 
             if (transaction.Receipt.Status == CoreModel.TransactionStatus.Succeeded)
             {
@@ -132,9 +156,23 @@ internal class ManifestProcessor : ITransactionMarkerProcessor, ITransactionScan
         return _manifestClasses.TryGetValue(stateVersion, out var mc) ? mc.ToArray() : Array.Empty<LedgerTransactionManifestClass>();
     }
 
-    private void AnalyzeManifestClasses(ToolkitModel.TransactionManifest toolkitManifest, long stateVersion)
+    private void AnalyzeManifestV1Classes(ToolkitModel.TransactionManifestV1 toolkitManifest, long stateVersion)
     {
-        var manifestSummary = toolkitManifest.Summary(_networkConfiguration.Id);
+        var manifestSummary = toolkitManifest.StaticallyAnalyze(_networkConfiguration.Id);
+
+        foreach (var manifestClass in manifestSummary.classification)
+        {
+            var mapped = manifestClass.ToModel();
+
+            _manifestClasses
+                .GetOrAdd(stateVersion, _ => new List<LedgerTransactionManifestClass>())
+                .Add(mapped);
+        }
+    }
+
+    private void AnalyzeManifestV2Classes(ToolkitModel.TransactionManifestV2 toolkitManifest, long stateVersion)
+    {
+        var manifestSummary = toolkitManifest.StaticallyAnalyze(_networkConfiguration.Id);
 
         foreach (var manifestClass in manifestSummary.classification)
         {